requirements.txt
amqp==1.4.9
anyjson==0.3.3
apache-libcloud==0.20.0
backports-abc==0.4
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.3.0.23
boto==2.38.0
celery==3.1.18
celery-with-redis==3.0
chardet==2.3.0
CherryPy==3.5.0
configobj==5.0.6
croniter==0.3.12
cssselect==0.9.1
cssutils==1.0
cycler==0.9.0
defer==1.0.4
Django==1.9.6
django-celery==3.2.1
django-crispy-forms==1.6.1
django-filter==1.0.1
django-formtools==2.0
django-pure-pagination==0.3.0
django-registration-redux==1.4
django-simple-captcha==0.4.6
djangorestframework==3.5.4
DjangoUeditor==1.8.143
dnspython==1.12.0
enum34==1.1.2
feedparser==5.1.3
futures==3.0.5
gitdb==0.6.4
GitPython==2.0.5
html5lib==0.999
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
Jinja2==2.8
kombu==3.0.37
lockfile==0.12.2
meld3==1.0.2
mongoengine==0.11.0
msgpack-python==0.4.8
netaddr==0.7.18
netifaces==0.10.4
numpy==1.11.0
olefile==0.44
Pillow==4.0.0
pycollada==0.4
pycrypto==2.6.1
pygame==1.9.3
Pygments==2.1
pyinotify==0.9.6
pymongo==3.4.0
PyMySQL==0.7.9
pyparsing==2.0.3
python-cloudfiles==1.7.10
python-dateutil==2.4.2
pytz==2016.10
pyxdg==0.25
PyYAML==3.11
pyzmq==15.2.0
redis==2.10.5
repoze.lru==0.6
requests==2.9.1
Routes==2.2
setuptools==20.10.1
simplejson==3.8.1
singledispatch==3.4.0.3
sis==0.1.5
six==1.10.0
smmap==0.9.0
supervisor==3.3.1
tornado==4.4.2
urllib3==1.13.1
uWSGI==2.0.15
vine==1.1.3
WebOb==1.5.1
wheel==0.29.0
xadmin==0.6.1